Lights and fabric compat
