+++ fabric/build.gradle
@@ -66,5 +66,5 @@ dependencies {
     modCompileOnly "foundry.imguimc:imguimc-fabric-${minecraft_version}:$imguimc_version"
 
     // Immersive portals compat
-    modCompileOnly "maven.modrinth:immersiveportals:v6.0.6-mc1.21.1"
+    modCompileOnly "maven.modrinth:immersiveportals:v${immptl_version}-mc${minecraft_version}"
 }
