+++ gradle.properties
@@ -25,6 +25,7 @@ imguimc_version=2.0.0
 lwjgl_version=3.3.3
 # There is no 0.8.12-alpha.3 for NeoForge, so just use 0.8.12-alpha.2
 sodium_version=0.8.12-alpha.2+mc1.21.1
+immptl_version=6.0.6
 
 # Fabric
 ## See https://fabricmc.net/develop for new versions
