+++ neoforge/build.gradle
@@ -69,5 +69,5 @@ dependencies {
     compileOnly "org.sinytra.forgified-fabric-api:forgified-fabric-api:0.104.0+2.0.19+1.21.1"
 
     // Immersive portals compat
-    compileOnly "maven.modrinth:immersive-portals-neoforge:6.0.7"
+    compileOnly "maven.modrinth:immersive-portals-neoforge:${immptl_version}"
 }
